Feat(EslintConfig): Add a new propmt for Eslint configuration.
Description:
The ESLint configuration file was previously updated automatically,
without giving users the option to choose their preferred indentation
style (e.g., tabs). This was not an ideal approach. In this update, a
new prompt has been added to give users the choice to update the ESLint
configuration file and specify their indentation preference.

Logs:
[Modified][Index]: Add a new optional prompt for eslint.
[Modified][FilesManager]: Check eslint condition before updating the
file.
[Modified][READNE]: Update readme file and add new options and featurs
to it.