Feat(FeaturesList): Add new flag to display list of available features.
Description:
Enhanced Nest Craft to include a new feature that displays a list of
available features for implementation in projects. Users can now view
all the features and options that can be added to a new or existing
project by using the `--list-features` flag.

Logs:
[Modified][ArgvManager]: Added support for the `--list-features` flag.
[Added][FeaturesList]: Add new functino to print available features
list.
[Modified][HelpDisplay]: Updated help documentation to include the usage
of the `--list-features` flag.